Divide generic functions of workflow.js
