+++ Cypress/cypress/pages/workflow/workflowValidation.js
@@ -1,0 +1,155 @@
+/// ************************** Verify and Commit Page **********************
+
+Cypress.Commands.add("verifyDetails", (name, description, schedule, min = 0) => {
+    cy.get("[data-cy=WorkflowName]").should("include.text", name); // Name validation
+    cy.get("[data-cy=WorkflowDescription]").should("have.text", description); // Description Validation
+    if (schedule == 0)
+        cy.get("[data-cy=schedule]").should("have.text", "Scheduling now"); // Schedule Validation
+    else if (schedule == 1)
+        cy.get("[data-cy=schedule]").should("have.text", `At ${min} minutes past the hour, between 12:00 AM and 11:59 PM`);
+    // cy.get("[data-cy=AgentName]").should("have.text", "Self-Agent");
+});
+
+/// ************************** Validate verdict of given workflow and agent **********************
+
+Cypress.Commands.add("validateVerdict", (workflowName, agentName, expectedVerdict, RScore, ExperimentsPassed, TotalExperiments) => {
+    cy.visit("/workflows");
+    cy.GraphqlWait("workflowListDetails", "listSchedules");
+    cy.get("[data-cy=runs]").click();
+    cy.wait("@listSchedules").its("response.statusCode").should("eq", 200);
+    cy.wait(1000);
+    cy.waitUntil(() =>
+        cy.get("table")
+            .find(`[data-cy=${workflowName}]`)
+            .find("[data-cy=WorkflowStatus]")
+            .then((status) => {
+                return status.text() != "Running" ? true : false;
+            }),
+        {
+            verbose: true,
+            interval: 500,
+            timeout: 600000,
+        }
+    );
+    cy.get('table')
+        .find(`[data-cy=${workflowName}]`)
+        .find("[data-cy=WorkflowStatus]")
+        .should("have.text", expectedVerdict);
+    cy.get('table')
+        .find(`[data-cy=${workflowName}]`)
+        .find("[data-cy=ResScore]")
+        .should("have.text", `Overall RR : ${RScore}%`);
+    cy.waitUntil(() =>
+        cy.get("table")
+            .find(`[data-cy=${workflowName}]`)
+            .find("[data-cy=ExperimentsPassed]")
+            .then((expPassed) => {
+                return expPassed.text() != "Experiments Passed : NA" ? true : false;
+            }),
+        {
+            verbose: true,
+            interval: 500,
+            timeout: 600000,
+        }
+    );
+    cy.get('table')
+        .find(`[data-cy=${workflowName}]`)
+        .find("[data-cy=ExperimentsPassed]")
+        .should("have.text", `Experiments Passed : ${ExperimentsPassed}/${TotalExperiments}`);
+});
+
+/// ************************** Validate dagre graph nodes **********************
+
+Cypress.Commands.add("validateGraphNodes", (graphNodesNameArray) => {
+    let nodes = [];
+    cy.waitUntil(() =>
+        cy.get("[data-cy=DagreGraphSvg]")
+            .find("text")
+            .each(($text) => {
+                nodes.push($text.text());
+            }).then(() => {
+                return graphNodesNameArray.every(val => nodes.includes(val));
+            }),
+        {
+            verbose: true,
+            interval: 500,
+            timeout: 60000,
+        }
+    );
+});
+
+/// ************************** Validate workflow info and stats **********************
+
+Cypress.Commands.add("validateWorkflowInfo", (workflowName, workflowNamespace, workflowSubject, agentName, regularity, nextRun) => {
+    cy.GraphqlWait("workflowListDetails", "SelectedWorkflowStats");
+    let workflowId = '';
+    let clusterId = '';
+    cy.wait("@recentRuns")
+        .then((res) => {
+            workflowId = res.response.body.data.ListWorkflow.workflows[0].workflow_id;
+            clusterId = res.response.body.data.ListWorkflow.workflows[0].cluster_id;
+            cy.get("[data-cy=infoWorkflowId]").should("have.text", workflowId);
+            cy.get("[data-cy=infoClusterId]").should("have.text", clusterId);
+        });
+    cy.get("[data-cy=statsWorkflowName]").should("have.text", workflowName);
+    cy.get("[data-cy=infoWorkflowName]").should("have.text", workflowName);
+    cy.get("[data-cy=infoWorkflowSubject]").should("have.text", workflowSubject);
+    cy.get("[data-cy=infoWorkflowNamespace]").should("have.text", workflowNamespace);
+    cy.get("[data-cy=infoAgentName]").should("have.text", agentName);
+
+    if (regularity === "Non cron workflow") {
+        cy.get("[data-cy=infoWorkflowRegularity]").should("have.text", `Regularity :${regularity}`);
+        cy.get("[data-cy=infoWorkflowNextRun]").should("have.text", `Next Run : ${nextRun}`);
+    }
+});
+
+///  Validate workflow/experiment stats radial chart, Passed ve Failed bar graph, RR Score chart 
+
+Cypress.Commands.add("validateStatsChart", () => {
+    cy.get("[data-cy=showStatsButton]").click();
+});
+
+/// ************************** Validate experiments table **********************
+
+Cypress.Commands.add("validateExperimentsTable", (experimentArray) => {
+    cy.waitUntil(() =>
+        cy.get("[data-cy=statsTable]")
+            .find("tr")
+            .eq(1)
+            .then(($tr) => {
+                return $tr.text() !== "No records available" ? true : false;
+            }),
+        {
+            verbose: true,
+            interval: 500,
+            timeout: 600000,
+        }
+    );
+    experimentArray.forEach((experiment, index) => {
+        cy.get("[data-cy=statsTable]")
+            .find("tr")
+            .eq(index + 1)
+            .then(($div) => {
+                //	Experiment Name
+                cy.wrap($div)
+                    .find("td")
+                    .eq(1)
+                    .should("have.text", experiment.experimentName);
+                // 	Experiment Verdict
+                cy.wrap($div)
+                    .find("td")
+                    .eq(2)
+                    .should("have.text", experiment.verdict);
+                // 	Weight of the test
+                cy.wrap($div)
+                    .find("td")
+                    .eq(3)
+                    .should("have.text", `${experiment.weightOfTest} Points`);
+                // 	Resulting Points
+                cy.wrap($div)
+                    .find("td")
+                    .eq(4)
+                    .should("have.text", `${experiment.resultingPoints} Points`);
+            });
+    });
+});
