+++ Cypress/cypress/support/pages.js
@@ -1,4 +1,9 @@
 import "../pages/getStarted";
 import "../pages/login";
-import "../pages/workflow";
+import "../pages/workflow/workflowClusterValidation";
+import "../pages/workflow/workflowCustom";
+import "../pages/workflow/workflowOperations";
+import "../pages/workflow/workflowPredefined";
+import "../pages/workflow/workflowRecurring";
+import "../pages/workflow/workflowValidation";
 import "../pages/settings";
